Move the sensuctl config command to its own pkg
