+++ cli/commands/config/set_organization.go
@@ -1,4 +1,4 @@
-package configure
+package config
 
 import (
 	"errors"
