+++ cli/commands/config/set_organization_test.go
@@ -1,4 +1,4 @@
-package configure
+package config
 
 import (
 	"errors"
